fix: clean up rebased theme verification
